SW-1449 Remove v1 accessions code from web-app codebase
- some refactor to keep shared constants
