--- src/components/seeds/lab/TableCellRenderer.tsx
@@ -1,53 +1,0 @@
-import { TableCell, Typography } from '@mui/material';
-import React from 'react';
-import strings from 'src/strings';
-import CellRenderer, { CellDateRenderer, TableRowType } from '../../common/table/TableCellRenderer';
-import { RendererProps } from '../../common/table/types';
-import { getDateDisplayValue } from '@terraware/web-components/utils';
-
-export default function LabCellRenderer(props: RendererProps<TableRowType>): JSX.Element {
-  const { column, row, value, index } = props;
-  if (column.key === 'seedsGerminated') {
-    return (
-      <CellRenderer
-        index={index}
-        column={column}
-        value={`${row[column.key]} ${strings.SEEDS_GERMINATED_LC}`}
-        row={row}
-      />
-    );
-  }
-  if (column.type === 'date' && typeof value === 'string' && value) {
-    const id = `row${index}-${column.key}`;
-    const date = new Date(value);
-    if (date > new Date()) {
-      return (
-        <TableCell id={id} align='left'>
-          <Typography component='p' variant='body1'>
-            {strings.SCHEDULED_FOR}
-          </Typography>
-          <Typography component='p' variant='body1'>
-            {getDateDisplayValue(value)}
-          </Typography>
-        </TableCell>
-      );
-    }
-
-    return <CellDateRenderer id={id} value={value} />;
-  }
-  if (column.key === 'seedsRemaining') {
-    return (
-      <CellRenderer
-        index={index}
-        column={column}
-        value={`${row.remainingQuantity?.quantity} ${row.remainingQuantity?.units}`}
-        row={row}
-      />
-    );
-  }
-  if (column.key === 'seedsSown') {
-    return <CellRenderer index={index} column={column} value={`${value} Seeds`} row={row} />;
-  }
-
-  return <CellRenderer {...props} />;
-}
