--- src/components/seeds/storage/LocationDropdown.tsx
@@ -1,89 +1,0 @@
-/* eslint-disable react-hooks/exhaustive-deps */
-import { Grid } from '@mui/material';
-import React, { useEffect, useState } from 'react';
-import { ConditionType, getLocations, StorageLocation } from 'src/api/seeds/locations';
-import Dropdown from 'src/components/common/Dropdown';
-import TextField from 'src/components/common/TextField';
-import strings from 'src/strings';
-import useDeviceInfo from 'src/utils/useDeviceInfo';
-
-interface LocationDropdownProps {
-  facilityId: number;
-  onChange: (id: string, value: unknown) => void;
-  storageLocation?: string;
-  storageCondition?: string;
-  isContributor: boolean;
-}
-
-export default function LocationDropdown(props: LocationDropdownProps): JSX.Element {
-  const { facilityId, onChange, storageLocation, storageCondition, isContributor } = props;
-  const [locations, setLocations] = useState<StorageLocation[]>([]);
-  const { isMobile } = useDeviceInfo();
-
-  const gridSize = () => {
-    if (isMobile) {
-      return 12;
-    }
-    return 4;
-  };
-
-  useEffect(() => {
-    const populateLocations = async () => {
-      const apiResponse = await getLocations(facilityId);
-      // TODO what if the request fails?
-      if (apiResponse !== null) {
-        setLocations(apiResponse);
-      }
-    };
-    populateLocations();
-  }, [facilityId]);
-
-  useEffect(() => {
-    if (storageLocation) {
-      onChange('storageCondition', getConditionValue(storageLocation) || '');
-    }
-  }, [storageLocation]);
-
-  const generateLocationsValues = locations?.map((location: StorageLocation) => {
-    return {
-      label: location.storageLocation,
-      value: location.storageLocation,
-    };
-  });
-
-  const getConditionValue = (locationSelected: string): ConditionType | undefined => {
-    const location = locations?.find((_location: StorageLocation) => {
-      return _location.storageLocation === locationSelected;
-    });
-
-    return location?.storageCondition;
-  };
-
-  const onStorageLocationChange = (id: string, value: string) => {
-    onChange(id, value);
-  };
-
-  return (
-    <>
-      <Grid item xs={gridSize()}>
-        <Dropdown
-          id='storageLocation'
-          label={strings.LOCATION}
-          selected={storageLocation}
-          values={generateLocationsValues}
-          onChange={onStorageLocationChange}
-          disabled={isContributor}
-        />
-      </Grid>
-      <Grid item xs={gridSize()}>
-        <TextField
-          id='storageCondition'
-          value={storageCondition}
-          onChange={onChange}
-          label={strings.CONDITION}
-          disabled={true}
-        />
-      </Grid>
-    </>
-  );
-}
