--- src/components/seeds/storage/StorageStartDate.tsx
@@ -1,73 +1,0 @@
-import { Grid } from '@mui/material';
-import moment from 'moment';
-import React, { useEffect, useState } from 'react';
-import { getDate } from 'src/api/clock';
-import DatePicker from 'src/components/common/DatePicker';
-import strings from 'src/strings';
-import useDeviceInfo from 'src/utils/useDeviceInfo';
-
-interface Props {
-  onChange: (id: string, value: string) => void;
-  refreshErrors: (newErrors: FieldError[]) => void;
-  storageDate?: string;
-  disabled?: boolean;
-}
-
-type FieldError = {
-  id: string;
-  msg: string;
-};
-export function StorageStartDate({ onChange, refreshErrors, storageDate, disabled }: Props): JSX.Element {
-  const [dateErrors, setDateErrors] = useState<FieldError[]>([]);
-  const [date, setDate] = useState<number>();
-  const { isMobile } = useDeviceInfo();
-
-  useEffect(() => {
-    const populateDate = async () => {
-      const response = await getDate();
-      setDate(response.serverTime ? response.serverTime : response.localTime);
-    };
-    populateDate();
-  }, []);
-
-  const onChangeDate = (id: string, value: unknown) => {
-    const newErrors = [...dateErrors];
-    const errorIndex = newErrors.findIndex((error) => error.id === id);
-    if (moment(value as string).isAfter(date)) {
-      if (errorIndex < 0) {
-        newErrors.push({
-          id,
-          msg: strings.NO_FUTURE_DATES,
-        });
-      }
-    } else {
-      if (errorIndex >= 0) {
-        newErrors.splice(errorIndex, 1);
-      }
-    }
-    setDateErrors(newErrors);
-    refreshErrors(newErrors);
-    onChange(id, value as string);
-  };
-
-  const getErrorText = (id: string) => {
-    const error = dateErrors.find((_error) => _error.id === id);
-
-    return error ? error.msg : '';
-  };
-
-  return (
-    <Grid item xs={isMobile ? 12 : 4}>
-      <DatePicker
-        id='storageStartDate'
-        value={storageDate}
-        onChange={onChangeDate}
-        label={strings.STARTING_ON}
-        aria-label='Starting on'
-        maxDate={moment(date).format('YYYY-MM-DD')}
-        errorText={getErrorText('storageStartDate')}
-        disabled={disabled}
-      />
-    </Grid>
-  );
-}
